microservices_backend/requirements.txt
flask
confluent-kafka
python-dotenv
requests
